# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CloneJira",
        "contact": {
            "email": "vvqua.2x@gmail.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "CloneJira"
        }
    ],
    "paths": {
...
        "/task/info/{id}": {
            "get": {
                "tags": [
                    "task"
                ],
                "summary": "Get Task Info By ID",
                "operationId": "05515104c881add4418445a8894fd09d",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/task/{id}": {
            "get": {
                "tags": [
                    "task"
                ],
                "summary": "Get List Task Info By Project ID",
                "operationId": "378898ea5e965eba98fddf42f2745706",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/task/add": {
            "post": {
                "tags": [
                    "task"
                ],
                "summary": "Create task",
                "operationId": "e724f4c6665a93319f369993c641867c",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_id": {
                                        "type": "integer"
                                    },
                                    "level_id": {
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "describes": {
                                        "type": "string"
                                    },
                                    "priority_id": {
                                        "type": "integer"
                                    },
                                    "datetimes": {
                                        "type": "string"
                                    },
                                    "user[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "id"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "project_id": 1,
                                    "level_id": 1,
                                    "category_id": 1,
                                    "name": "hi KenDzz",
                                    "describes": "hi KenDzz",
                                    "priority_id": 1,
                                    "datetimes": "3/07/2024 04:00 PM - 3/23/2024 12:00 AM",
                                    "user[]": 2
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/task/delete/{id}": {
            "delete": {
                "tags": [
                    "task"
                ],
                "summary": "Delete task",
                "operationId": "56efdb95cdbea2da3851e19b5c77ac82",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/task/update": {
            "post": {
                "tags": [
                    "task"
                ],
                "summary": "Update task",
                "operationId": "ff29fab0746adc0b1803dad52877008c",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "level_id": {
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "describes": {
                                        "type": "string"
                                    },
                                    "priority_id": {
                                        "type": "integer"
                                    },
                                    "datetimes": {
                                        "type": "string"
                                    },
                                    "user[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "id"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "level_id": 1,
                                    "category_id": 1,
                                    "name": "hi KenDzz",
                                    "describes": "hi KenDzz",
                                    "priority_id": 1,
                                    "datetimes": "3/07/2024 04:00 PM - 3/23/2024 12:00 AM",
                                    "user[]": 2
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/task/level/update": {
            "post": {
                "tags": [
                    "task"
                ],
                "summary": "Update level task",
                "operationId": "1f8a888ac751c156f1ff862820e90ffb",
                "parameters": [
                    {
                        "name": "X-localization",
                        "in": "header",
                        "description": "Set language parameter (Example: 'vi' or 'en'))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "level_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "level_id": 1,
                                    "type": 0
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}